compositor: split config header
