+++ compositor/config/config.c
@@ -48,6 +48,11 @@
 #include "bindings.h"
 #include "config.h"
 
+struct tw_config_obj {
+	char name[32];
+	void *data;
+};
+
 /******************************************************************************
  * internal helpers
  *****************************************************************************/
