+++ compositor/config/config.h
@@ -1,5 +1,5 @@
 /*
- * config_internal.h - taiwins config shared header
+ * config.h - taiwins config shared header
  *
  * Copyright (c) 2019 Xichen Zhou
  *
@@ -32,6 +32,8 @@
 
 #include <taiwins/engine.h>
 #include "xdg.h"
+#include "config_types.h"
+#include "config_bindings.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -44,38 +46,6 @@ extern "C" {
 enum tw_config_type {
 	TW_CONFIG_TYPE_LUA,
 	TW_CONFIG_TYPE_DBUS,
-};
-
-enum tw_builtin_binding_t {
-	TW_QUIT_BINDING = 0,
-	TW_CLOSE_APP_BINDING,
-	TW_RELOAD_CONFIG_BINDING,
-	//QUIT taiwins, rerun configuration
-	//console
-	TW_OPEN_CONSOLE_BINDING,
-	//shell
-	TW_ZOOM_AXIS_BINDING,
-	TW_ALPHA_AXIS_BINDING,
-	//views
-	TW_MOVE_PRESS_BINDING,
-	TW_FOCUS_PRESS_BINDING,
-	TW_RESIZE_ON_LEFT_BINDING,
-	TW_RESIZE_ON_RIGHT_BINDING,
-	TW_RESIZE_ON_UP_BINDING,
-	TW_RESIZE_ON_DOWN_BINDING,
-	//workspace
-	TW_SWITCH_WS_LEFT_BINDING,
-	TW_SWITCH_WS_RIGHT_BINDING,
-	TW_SWITCH_WS_RECENT_BINDING,
-	TW_TOGGLE_FLOATING_BINDING,
-	TW_TOGGLE_VERTICAL_BINDING,
-	TW_VSPLIT_WS_BINDING,
-	TW_HSPLIT_WS_BINDING,
-	TW_MERGE_BINDING,
-	//view cycling
-	TW_NEXT_VIEW_BINDING,
-	//sizeof
-	TW_BUILTIN_BINDING_SIZE
 };
 
 /**
@@ -93,41 +63,6 @@ enum tw_config_enable_global {
 	TW_CONFIG_GLOBAL_XWAYLAND = (1 << 5),
 	TW_CONFIG_GLOBAL_DESKTOP = (1 << 6),
 };
-
-struct tw_config_obj {
-	char name[32];
-	void *data;
-};
-
-typedef struct {
-	int rotate;
-	bool flip;
-	enum wl_output_transform t;
-} tw_config_transform_t;
-
-typedef OPTION(enum wl_output_transform, transform) pending_transform_t;
-typedef OPTION(enum tw_layout_type, layout) pending_layout_t;
-typedef OPTION(bool, enable) pending_boolean_t;
-typedef OPTION(enum taiwins_shell_task_switch_effect, eff) pending_effect_t;
-typedef OPTION(enum taiwins_shell_panel_pos, pos) pending_panel_pos_t;
-typedef OPTION(int32_t, val) pending_intval_t;
-typedef OPTION(uint32_t, uval) pending_uintval_t;
-typedef OPTION(struct tw_theme *, theme) pending_theme_t;
-
-#define SET_PENDING(ptr, name, value)                                   \
-	({ \
-		(ptr)->name = value; \
-		(ptr)->valid = true; \
-	})
-
-struct tw_config_output {
-	char name[24];
-	pending_intval_t scale, posx, posy;
-	pending_uintval_t width, height;
-	pending_transform_t transform;
-	pending_boolean_t enabled;
-};
-
 
 struct tw_config_table {
 	bool dirty;
@@ -200,15 +135,7 @@ tw_config_register_object(struct tw_config *config,
 void *
 tw_config_request_object(struct tw_config *config,
                          const char *name);
-void
-tw_config_default_bindings(struct tw_config *c);
 
-bool
-tw_config_install_bindings(struct tw_config *c, struct tw_bindings *root);
-
-bool
-parse_one_press(const char *str, const enum tw_binding_type type,
-                uint32_t *mod, uint32_t *code);
 const char *
 tw_config_retrieve_error(struct tw_config *);
