+++ compositor/config/config_bindings.h
@@ -1,0 +1,80 @@
+/*
+ * config_bindings.h - taiwins config bindings implementation
+ *
+ * Copyright (c) 2020 Xichen Zhou
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, write to the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ */
+
+#ifndef TW_CONFIG_BINDINGS_H
+#define TW_CONFIG_BINDINGS_H
+
+#include <stdbool.h>
+#include "bindings.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+struct tw_config;
+
+enum tw_builtin_binding_t {
+	TW_QUIT_BINDING = 0,
+	TW_CLOSE_APP_BINDING,
+	TW_RELOAD_CONFIG_BINDING,
+	//QUIT taiwins, rerun configuration
+	//console
+	TW_OPEN_CONSOLE_BINDING,
+	//shell
+	TW_ZOOM_AXIS_BINDING,
+	TW_ALPHA_AXIS_BINDING,
+	//views
+	TW_MOVE_PRESS_BINDING,
+	TW_FOCUS_PRESS_BINDING,
+	TW_RESIZE_ON_LEFT_BINDING,
+	TW_RESIZE_ON_RIGHT_BINDING,
+	TW_RESIZE_ON_UP_BINDING,
+	TW_RESIZE_ON_DOWN_BINDING,
+	//workspace
+	TW_SWITCH_WS_LEFT_BINDING,
+	TW_SWITCH_WS_RIGHT_BINDING,
+	TW_SWITCH_WS_RECENT_BINDING,
+	TW_TOGGLE_FLOATING_BINDING,
+	TW_TOGGLE_VERTICAL_BINDING,
+	TW_VSPLIT_WS_BINDING,
+	TW_HSPLIT_WS_BINDING,
+	TW_MERGE_BINDING,
+	//view cycling
+	TW_NEXT_VIEW_BINDING,
+	//sizeof
+	TW_BUILTIN_BINDING_SIZE
+};
+
+void
+tw_config_default_bindings(struct tw_config *c);
+
+bool
+tw_config_install_bindings(struct tw_config *c, struct tw_bindings *root);
+
+bool
+parse_one_press(const char *str, const enum tw_binding_type type,
+                uint32_t *mod, uint32_t *code);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* EOF */
